initial commit
the state of the project is as the end of birch second video (https://www.youtube.com/watch?v=jzasDqPmtPI), with the difference that I am building pong instead of his game.

this has SLD2 library initialization, a game loop and drawing on the screen with png. and a counter!